# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TrackFlow API — digital tracking governance platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workspaces:
    get:
      operationId: listWorkspaces
      tags: [workspaces]
      summary: List all workspaces for the current user
      responses:
        "200":
          description: List of workspaces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workspace"
    post:
      operationId: createWorkspace
      tags: [workspaces]
      summary: Create a new workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
  /workspaces/{workspaceId}:
    get:
      operationId: getWorkspace
      tags: [workspaces]
      summary: Get a workspace by ID
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
        "404":
          description: Not found
    patch:
      operationId: updateWorkspace
      tags: [workspaces]
      summary: Update a workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceUpdate"
      responses:
        "200":
          description: Updated workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"

  # ─── Dashboard ────────────────────────────────────────────────────────────
...
  /workspaces/{workspaceId}/webhooks:
    get:
      operationId: listWebhookEndpoints
      tags: [webhooks]
      summary: List all webhook endpoints for a workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of webhook endpoints
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WebhookEndpoint"
    post:
      operationId: createWebhookEndpoint
      tags: [webhooks]
      summary: Create a new webhook endpoint
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookEndpointInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookEndpoint"
  /workspaces/{workspaceId}/webhooks/{endpointId}:
    get:
      operationId: getWebhookEndpoint
      tags: [webhooks]
      summary: Get a webhook endpoint with its recent events
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
        - name: endpointId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Webhook endpoint detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookEndpointDetail"
    delete:
      operationId: deleteWebhookEndpoint
      tags: [webhooks]
      summary: Delete a webhook endpoint
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
        - name: endpointId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /workspaces/{workspaceId}/webhooks/{endpointId}/events:
    get:
      operationId: listWebhookEvents
      tags: [webhooks]
      summary: List events received by a webhook endpoint
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
        - name: endpointId
          in: path
          required: true
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of webhook events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WebhookEvent"
  /workspaces/{workspaceId}/webhooks/{endpointId}/events/{eventId}/resend:
    post:
      operationId: resendWebhookEvent
      tags: [webhooks]
      summary: Resend a webhook event
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
        - name: endpointId
          in: path
          required: true
          schema:
            type: integer
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resend result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookEvent"

  # ─── Events / Timeline ────────────────────────────────────────────────────
...
components:
  schemas:
...
    Workspace:
      type: object
      required: [id, name, slug, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        logoUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    WorkspaceInput:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
          minLength: 1
        slug:
          type: string
          minLength: 1
        logoUrl:
          type: string
    WorkspaceUpdate:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        logoUrl:
          type: string

    # ─── Dashboard ─────────────────────────────────────────────────────────
...
    WebhookEndpoint:
      type: object
      required: [id, workspaceId, name, slug, url, createdAt, totalEvents]
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        name:
          type: string
        slug:
          type: string
        url:
          type: string
        description:
          type: ["string", "null"]
        totalEvents:
          type: integer
        lastEventAt:
          type: ["string", "null"]
        createdAt:
          type: string
    WebhookEndpointInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
    WebhookEndpointDetail:
      type: object
      required: [id, workspaceId, name, slug, url, createdAt, totalEvents, recentEvents]
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        name:
          type: string
        slug:
          type: string
        url:
          type: string
        description:
          type: ["string", "null"]
        totalEvents:
          type: integer
        lastEventAt:
          type: ["string", "null"]
        createdAt:
          type: string
        recentEvents:
          type: array
          items:
            $ref: "#/components/schemas/WebhookEvent"
    WebhookEvent:
      type: object
      required: [id, endpointId, timestamp, method, statusCode, ip]
      properties:
        id:
          type: integer
        endpointId:
          type: integer
        timestamp:
          type: string
        method:
          type: string
        headers:
          type: object
          additionalProperties:
            type: string
        payload:
          type: object
          additionalProperties: true
        statusCode:
          type: integer
        ip:
          type: ["string", "null"]
        origin:
          type: ["string", "null"]
        status:
          type: string
          enum: [received, processed, failed]

    # ─── Events ────────────────────────────────────────────────────────────
...