# lib/api-zod/src/generated/types/webhookEndpointDetail.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * TrackFlow API — digital tracking governance platform
 * OpenAPI spec version: 0.1.0
 */
import type { WebhookEvent } from './webhookEvent';

export interface WebhookEndpointDetail {
  id: number;
  workspaceId: number;
  name: string;
  slug: string;
  url: string;
  /** @nullable */
  description?: string | null;
  totalEvents: number;
  /** @nullable */
  lastEventAt?: string | null;
  createdAt: string;
  recentEvents: WebhookEvent[];
}