# lib/api-zod/src/generated/types/webhookEndpointInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * TrackFlow API — digital tracking governance platform
 * OpenAPI spec version: 0.1.0
 */

export interface WebhookEndpointInput {
  /** @minLength 1 */
  name: string;
  description?: string;
}